feat(components/topbar/user): refactor, and extract fixtures
